Add error handling to check directus url

+++ server/api/images/[id].delete.ts
@@ -1,13 +1,23 @@
 import { createDirectus, rest, deleteFile } from "@directus/sdk";
 
 export default defineEventHandler(async (event) => {
-  await requireUserSession(event);
-  const directus = createDirectus("https://directus.hoachnt.com").with(rest());
-  const id = getRouterParam(event, "id");
+	await requireUserSession(event);
 
-  if (id === undefined) return;
+	if (typeof process.env.NUXT_PUBLIC_DIRECTUS_URL !== "string") {
+		throw createError({
+			statusCode: 500,
+			statusMessage: "DIRECTUS_URL is not defined",
+		});
+	}
 
-  const result = await directus.request(deleteFile(id));
+	const directus = createDirectus(process.env.NUXT_PUBLIC_DIRECTUS_URL).with(
+		rest()
+	);
+	const id = getRouterParam(event, "id");
 
-  return result;
+	if (id === undefined) return;
+
+	const result = await directus.request(deleteFile(id));
+
+	return result;
 });
